requirements.txt
pyspark-3.5.0
boto3